Beta-2 update
Moved App Setting Tab button code to AppSettingsTab.cs
Added 2x2 button option when there are 4 buttons on AppSettings
Cleared helper log transitions when no reachable transitions are found in log
Updated readme to reflect v2 changes
